pyproject.toml
[tool.poetry]
name = "weather-mcp"
version = "0.1.0"
description = "Weather application using Model Context Protocol"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"   # 最新版！
uvicorn = "^0.23.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.8.0"
pydantic = "^2.7.2"
mcp = { git = "https://github.com/modelcontextprotocol/python-sdk.git" }
typer = "^0.15.2"
rich = "^14.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
